Restructure linfa-datasets to fix warnings
